+++ packages/espressocash_app/lib/core/link_payments.dart
@@ -20,7 +20,10 @@ abstract class LinkPayments with _$LinkPayments {
   const LinkPayments._();
 
   static LinkPayments? tryParse(Uri link) {
-    if (link.scheme != 'https' || link.host != espressoCashLinkDomain) {
+    if ((link.scheme != 'https' ||
+            link.host != espressoCashLinkDomain ||
+            link.queryParameters['t'] != 'link') &&
+        link.scheme != espressoCashLinkProtocol) {
       return null;
     }
 
@@ -38,6 +41,16 @@ abstract class LinkPayments with _$LinkPayments {
         host: '',
         path: '',
         queryParameters: <String, String>{
+          'k': key,
+        },
+      );
+
+  Uri toShareableLink() => Uri(
+        scheme: 'https',
+        host: espressoCashLinkDomain,
+        path: '',
+        queryParameters: <String, String>{
+          't': 'link',
           'k': key,
         },
       );