+++ packages/espressocash_app/lib/features/activities/data/pending_activities_repository.dart
@@ -7,7 +7,7 @@ import 'package:rxdart/rxdart.dart';
 import '../../../core/tokens/token_list.dart';
 import '../../../data/db/db.dart';
 import '../../outgoing_direct_payments/data/repository.dart';
-import '../../outgoing_split_key_payments/data/repository.dart';
+import '../../outgoing_link_payments/data/repository.dart';
 import '../../payment_request/data/repository.dart';
 import '../../ramp/data/on_ramp_order_service.dart';
 import '../../swap/data/swap_repository.dart';
@@ -33,20 +33,20 @@ class PendingActivitiesRepository {
       );
     final odp = _db.select(_db.oDPRows)
       ..where((tbl) => tbl.status.equalsValue(ODPStatusDto.success).not());
-    final oskp = _db.select(_db.oSKPRows)
-      ..where((tbl) => tbl.status.equalsValue(OSKPStatusDto.withdrawn).not())
-      ..where((tbl) => tbl.status.equalsValue(OSKPStatusDto.canceled).not());
     final swap = _db.select(_db.swapRows)
       ..where((tbl) => tbl.status.equalsValue(SwapStatusDto.success).not());
+    final olp = _db.select(_db.oLPRows)
+      ..where((tbl) => tbl.status.equalsValue(OLPStatusDto.withdrawn).not())
+      ..where((tbl) => tbl.status.equalsValue(OLPStatusDto.canceled).not());
 
     final oprStream =
         opr.watch().map((rows) => rows.map((r) => r.toActivity()));
     final odpStream =
         odp.watch().map((rows) => rows.map((r) => r.toActivity(_tokens)));
-    final oskpStream =
-        oskp.watch().map((rows) => rows.map((r) => r.toActivity(_tokens)));
     final swapStream =
         swap.watch().map((rows) => rows.map((r) => r.toActivity(_tokens)));
+    final olpStream =
+        olp.watch().map((rows) => rows.map((r) => r.toActivity(_tokens)));
 
     final onRampStream = _onRampOrderService.watchPending().map(
           (rows) =>
@@ -57,9 +57,9 @@ class PendingActivitiesRepository {
       [
         oprStream,
         odpStream,
-        oskpStream,
         swapStream,
         onRampStream,
+        olpStream,
       ],
       (values) => values
           .expand(identity)