+++ packages/espressocash_app/lib/features/activities/widgets/transaction_item.dart
@@ -5,8 +5,8 @@ import '../data/transaction_repository.dart';
 import '../models/transaction.dart';
 import 'common_tile.dart';
 import 'odp_tile.dart';
+import 'olp_tile.dart';
 import 'on_ramp_tile.dart';
-import 'oskp_tile.dart';
 import 'payment_request_tile.dart';
 import 'swap_tile.dart';
 
@@ -52,7 +52,7 @@ class _TransactionItemState extends State<TransactionItem> {
                 key: ValueKey(p.id),
                 activity: p,
               ),
-              outgoingSplitKeyPayment: (p) => OSKPTile(
+              outgoingLinkPayment: (p) => OLPTile(
                 key: ValueKey(p.id),
                 activity: p,
               ),