+++ packages/espressocash_app/lib/features/investments/widgets/home_carousel.dart
@@ -37,17 +37,7 @@ class _HomeCarouselWidgetState extends State<HomeCarouselWidget> {
           text: context.l10n.carouselSendMoneyBtn,
           onPressed: () => context.router.navigate(WalletFlowScreen.route()),
         ),
-        backgroundImage: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Expanded(
-              child: Assets.images.carousel1.image(),
-            ),
-            Expanded(
-              child: Assets.images.carousel15.image(),
-            ),
-          ],
-        ),
+        backgroundImage: Assets.images.carousel1.image(),
       ),
       _Item(
         title: context.l10n.carousel2Title,
@@ -83,7 +73,7 @@ class _HomeCarouselWidgetState extends State<HomeCarouselWidget> {
                 count: items.length,
                 effect: const ExpandingDotsEffect(
                   activeDotColor: CpColors.primaryColor,
-                  dotColor: Colors.white,
+                  dotColor: Colors.black,
                   dotHeight: 7,
                   dotWidth: 25,
                   expansionFactor: 2.4,