+++ packages/espressocash_app/lib/features/wallet_flow/screens/wallet_flow_screen.dart
@@ -16,9 +16,9 @@ import '../../outgoing_direct_payments/screens/odp_confirmation_screen.dart';
 import '../../outgoing_direct_payments/screens/odp_details_screen.dart';
 import '../../outgoing_direct_payments/screens/odp_input_screen.dart';
 import '../../outgoing_direct_payments/widgets/extensions.dart';
-import '../../outgoing_split_key_payments/screens/oskp_confirmation_screen.dart';
-import '../../outgoing_split_key_payments/screens/oskp_screen.dart';
-import '../../outgoing_split_key_payments/widgets/extensions.dart';
+import '../../outgoing_link_payments/screens/olp_confirmation_screen.dart';
+import '../../outgoing_link_payments/screens/olp_screen.dart';
+import '../../outgoing_link_payments/widgets/extensions.dart';
 import '../../payment_request/screens/link_details_flow_screen.dart';
 import '../../payment_request/widgets/extensions.dart';
 import '../../qr_scanner/widgets/build_context_ext.dart';
@@ -105,7 +105,7 @@ class _State extends State<WalletFlowScreen> {
       PayFlowScreen.route(
         onLink: () {
           context.router.push(
-            OSKPConfirmationScreen.route(
+            OLPConfirmationScreen.route(
               tokenAmount: cryptoAmount,
               // TODO(KB): do not hardcode
               fee: Amount.fromDecimal(
@@ -113,13 +113,13 @@ class _State extends State<WalletFlowScreen> {
                 currency: Currency.usdc,
               ),
               onSubmit: () async {
-                final id = await context.createOSKP(amount: cryptoAmount);
+                final id = await context.createOLP(amount: cryptoAmount);
                 if (!mounted) return;
 
                 await context.router.pop();
                 if (!mounted) return;
 
-                await context.router.replace(OSKPScreen.route(id: id));
+                await context.router.replace(OLPScreen.route(id: id));
                 if (!mounted) return;
 
                 setState(() => _fiatAmount = _fiatAmount.copyWith(value: 0));