+++ packages/espressocash_app/lib/ui/usdc_info.dart
@@ -13,18 +13,18 @@ class UsdcInfoWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => Padding(
-        padding: EdgeInsets.symmetric(
-          vertical: 8,
-          horizontal: isSmall ? 20 : 60,
-        ),
+        padding: const EdgeInsets.symmetric(horizontal: 45),
         child: CpInfoWidget(
           message: Text(
             context.l10n.usdcExplanation,
-            style: TextStyle(fontSize: isSmall ? 12 : 14.5),
-            maxLines: 2,
+            style: TextStyle(
+              fontSize: isSmall ? 12 : 14.5,
+              fontWeight: FontWeight.w500,
+            ),
+            maxLines: 3,
           ),
           variant: CpInfoVariant.black,
-          padding: EdgeInsets.all(isSmall ? 12 : 20),
+          padding: EdgeInsets.all(isSmall ? 14 : 18),
         ),
       );
 }